fixx

+++ NMTL-NguyenXuanHung-PS50057-main/lab 2/Bai 2/Bai2_Tinhdiem.c
@@ -2,19 +2,19 @@
 #include <stdlib.h>
 
 int main() {
-    int diemToan, diemLy, diemHoa, diemTB;
+    float diemToan, diemLy, diemHoa, diemTB;
 
     printf("Nhap diem Toan: ");
-    scanf("%d", &diemToan);
+    scanf("%f", &diemToan);
     printf("Nhap diem Ly: ");
-    scanf("%d", &diemLy);
+    scanf("%f", &diemLy);
     printf("Nhap diem Hoa: ");  
-    scanf("%d", &diemHoa);
+    scanf("%f", &diemHoa);
 
     diemTB = (diemToan * 3 + diemLy * 2 + diemHoa) / 6;
 
     printf("===KET QUA===\n");
-    printf("Diem trung binh: %d\n", diemTB);
+    printf("Diem trung binh: %.2f\n", diemTB);
 
     system("pause");
     return 0;
